+++ laravel/app/Http/Requests/UserPostDeleteRequest.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+
+class UserPostDeleteRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'password' => [
+				'required',
+				function ($attribute, $value, $fail) {
+					if (!Hash::check($value, Auth::user()->password)) {
+						$fail('パスワードが間違っています。');
+					}
+				},
+			],
+        ];
+    }
+
+    public function messages()
+    {
+        return [
+			'password.required'   => 'パスワードを入力して下さい。',
+        ];
+    }
+}
